Feat. 푸시 알림 및 자동로그인 기능 구현

+++ app/src/main/java/com/example/meetup/model/GetHeartListResponseModel.kt
@@ -17,7 +17,7 @@ data class GetHeartListResponseModelStore(
     val name: String,
     val description: String,
     val koreanYn: Boolean,
-    val avgRate: Long,
+    val avgRate: Double,
     val minOrderAmount: Long,
     val images: ArrayList<String>,
 )
