+++ app/src/main/java/com/example/meetup/model/request/OrderFoodRequestModel.kt
@@ -2,7 +2,7 @@ package com.example.meetup.model.request
 
 data class OrderFoodRequestModel(
     val store: Long,
-    val totalPrice: Long,
+    val totalPrice: Double,
     val totalCount: Long,
     val selectedOption: String,
     val orderFoodDetailList: List<OrderFoodDetailList>,
@@ -12,5 +12,5 @@ data class OrderFoodDetailList(
     val foodId: Long,
     val foodOptionId: Long,
     val orderCount: Long,
-    val orderAmount: Long,
+    val orderAmount: Double,
 )
